+++ contracts/extensions/SemaphoreWhistleblowing.sol
@@ -84,7 +84,7 @@ contract SemaphoreWhistleblowing is ISemaphoreWhistleblowing, SemaphoreCore, Sem
         bytes calldata roots,
         uint256[8] calldata proof
     ) public override onlyEditor(entityId) {
-        uint root = getRoot(entityId);
+        uint256 root = getRoot(entityId);
         uint8 depth = getDepth(entityId);
         uint8 maxEdges = getMaxEdges(entityId);
         SemaphoreVerifier verifier = verifiers[depth];
